Update pre_submission.py
Method run had 82 statements while pylint suggest not to have more than 50.
I extracted some small local methods, adding structure and making the method shorter.

Please note that the end of the method was not reachable (and still so) due to a break, which might be a bug.
I added a TODO there to mark that.